Move consensus convenience functions to a separate module
